SPA_Table_2/requirements.txt
mysql-connector-python==8.0.25
protobuf==3.17.0
six==1.16.0
